Feat: 회원가입/로그인 (#11)
* #1 feat: 디바이스 존재 여부 확인 controller,service,repository 구현

* #1 Feat: 네이버 클라우드 기반 문자메세지 전송 기능 구현

* #1 Feat: 디바이스 존재 여부 기능 구현

* #1 Refactor: openapi와 통신할 http 통신 템플릿을 RestTemplate으로 변경

* #1 Feat: 휴대폰 번호로 인증코드 전송 후 코드 세선에 저장, 이후 입력값이 발송코드와 일치 여부에 따라 결과값 return

* #1 Fix: MySQL 예약어인ㅇ 컬럼 수정

* #1 Feat: 회원가입 요청 후 유효성 검사와 db 저장 구현. password,pattern은 salt로 암호화

* #1 Feat: 회원가입, 로그인(password) 구현 성공 - JWT이용해서 access,refresh token 발급. 로그인 시 deviceId와 password 입력

* #6 Feat: 환전 요청을 받고 환전처리 초기 구현

* #1 Feat: spring security+jwt(access_token, refresh_token)을 이용한 로그인과 기타기능 인가 구현

* #1 Docs: 불필요한 주석 및 쓰지않는 import 삭제

* #1 Fix: refresh token으로 access token 재발급 안되는 문제 해결

* #1 Fix: Spring security - swagger 허용

* #1 Fix: access,refresh token body에서 제외하고 header에만 추가, 인가 성공 코드 추가

* #1 Feat: ApiReponse 추가

* #1 Refactor: 코드 컨벤션 적용, 로그인 테스트용 API 추가

* Refactor: ErrorCode의 status를 상태코드에서 enum 이름으로 변경

* Refactor: GlobalExceptionHandler 통일

* Refactor: @RequiredConstructor 통일,쓰지않는 hashmap 삭제

+++ src/main/java/com/example/travelhana/Domain/ExchangeHistory.java
@@ -22,13 +22,13 @@ public class ExchangeHistory {
 	private Long id;
 
 	@Column
-	private Long userId;
+	private int userId;
 
 	@Column
-	private Long keyId;
+	private int keyId;
 
 	@Column
-	private Long accountId;
+	private int accountId;
 
 	@Column
 	private Long won;
